benchmarks/requirements.txt
numpy
pandas
matplotlib
psutil
tqdm
pinecone-client
psycopg2-binary
pgvector
